+++ auth.js
@@ -1,6 +1,6 @@
 // @flow
 import { NativeModules, NativeEventEmitter } from 'react-native';
-import type { User, AuthCredential } from './types';
+import type { User, AuthCredential, AuthModule } from './types';
 const NativeFirebaseBridgeAuth = NativeModules.FirebaseBridgeAuth;
 const {
     FirebaseBridgeFacebookAuthProvider,
@@ -21,44 +21,14 @@ async function signInWithCredential(credential:AuthCredential|Promise<AuthCreden
 const signInAnonymously:() => Promise<User> =
     NativeFirebaseBridgeAuth.signInAnonymously;
 
-const currentUser:() => Promise<User> = 
-    NativeFirebaseBridgeAuth.currentUser;
+const getCurrentUser:() => Promise<User> =
+    NativeFirebaseBridgeAuth.getCurrentUser;
 
 const signOut:() => Promise<null> =
     NativeFirebaseBridgeAuth.signOut;
-type AuthStateListener = (user:User) => void;
+type AuthStateListener = (user:?User) => void;
 
 const authStateDidChangeListeners:Array<AuthStateListener> = [];
-
-let authUser:User;
-const authEmitter = new NativeEventEmitter(NativeFirebaseBridgeAuth);
-const subscription = authEmitter.addListener(
-    'authStateDidChange',
-    (user:User) => {
-        authUser = user;
-        authStateDidChangeListeners.forEach(cb => cb(user));
-    }
-);
-
-let authStateDidChangeListenerRegistered = false;
-function addAuthStateDidChangeListener(cb:AuthStateListener) : () => void {
-    authStateDidChangeListeners.push(cb);
-    if (!authStateDidChangeListenerRegistered) {
-        NativeFirebaseBridgeAuth.addAuthStateDidChangeListener();
-        authStateDidChangeListenerRegistered = true;
-    } else {
-        cb(authUser);
-    }
-    return () => {
-        const index = authStateDidChangeListeners.indexOf(cb);
-        if (index === -1) {
-            console.warn( // eslint-disable-line
-                'Callback not found for authStateDidChangeListener');
-            return;
-        }
-        authStateDidChangeListeners.splice(index, 1);
-    };
-}
 
 const FacebookAuthProvider = {
     credential(token:string) : Promise<AuthCredential> {
@@ -84,28 +54,70 @@ const GithubAuthProvider = {
     },
 };
 
-export default {
-    addAuthStateDidChangeListener,
+const Auth:AuthModule = {
+    currentUser: null,
+    onAuthStateChanged, // eslint-disable-line
+    addAuthStateDidChangeListener, // eslint-disable-line
     createUserWithEmail,
     signInWithEmail,
     signInAnonymously,
     signInWithCredential,
     signOut,
-    currentUser,
+    getCurrentUser,
     FacebookAuthProvider,
     GithubAuthProvider,
     TwitterAuthProvider,
     GoogleAuthProvider,
 };
+const authEmitter = new NativeEventEmitter(NativeFirebaseBridgeAuth);
+const subscription = authEmitter.addListener(
+    'authStateDidChange',
+    (user?:User) => {
+        Auth.currentUser = user;
+        authStateDidChangeListeners.forEach(cb => cb(user));
+    }
+);
 
+// Always add listener immediately so currentUser is set as soon as possible
+function onAuthStateChanged(cb:AuthStateListener) : () => void {
+    authStateDidChangeListeners.push(cb);
+    cb(Auth.currentUser);
+    return () => {
+        const index = authStateDidChangeListeners.indexOf(cb);
+        if (index === -1) {
+            console.warn( // eslint-disable-line
+                'Callback not found for authStateDidChangeListener');
+            return;
+        }
+        authStateDidChangeListeners.splice(index, 1);
+    };
+}
+
+function addAuthStateDidChangeListener(cb:(payload:?{ user: User }) => void) : () => void {
+    console.warn( // eslint-disable-line
+        'addAuthStateDidChangeListener() is deprecated; use onAuthStateChanged. ' +
+        'Callback passed to onAuthStateChanged will receive null or the user rather ' +
+        "than null or an object with a 'user' key");
+    return onAuthStateChanged(user => {
+        if (user) {
+            cb({ user });
+        } else {
+            cb(null);
+        }
+    });
+}
+
+
+export default Auth;
 export {
     addAuthStateDidChangeListener,
+    onAuthStateChanged,
     createUserWithEmail,
     signInWithEmail,
     signInAnonymously,
     signInWithCredential,
     signOut,
-    currentUser,
+    getCurrentUser,
     FacebookAuthProvider,
     GithubAuthProvider,
     TwitterAuthProvider,