+++ types.js
@@ -8,7 +8,7 @@ export interface DataSnapshot {
     child(path: string) : DataSnapshot;
     exists() : Promise<boolean>;
     exportVal() : Promise<any>;
-    forEach(cb:(snapshot:DataSnapshot) => Promise) : Promise;
+    forEach(cb:(snapshot:DataSnapshot) => Promise<?boolean>) : Promise<void>;
     getPriority() : Promise<Priority>;
     hasChild(path:string) : Promise<boolean>;
     hasChildren() : Promise<boolean>;
@@ -21,8 +21,8 @@ export interface Query {
     equalTo(value:number|string|boolean|null, key:?string) : Query;
     limitToFirst(limit:number) : Query;
     limitToLast(limit:number) : Query;
-    on(eventType:EventType, cb:((snapshot:DataSnapshot) => Promise)) : () => void;
-    once(eventType:EventType, cb:((snapshot:DataSnapshot) => Promise)) : () => void;
+    on(eventType:EventType, cb:((snapshot:DataSnapshot) => Promise<void>)) : () => void;
+    once(eventType:EventType, cb:((snapshot:DataSnapshot) => Promise<void>)) : () => void;
     orderByChild(path:string) : Query;
     orderByKey() : Query;
     orderByPriority() : Query;
@@ -35,11 +35,10 @@ export type DatabaseReference = Query & {
     key(): Promise<string>;
     child(pathString:string) : DatabaseReference;
     push() : DatabaseReference;
-    setValue(value:any) : Promise;
-    setValueWithPriority(value:any, priority:Priority) : Promise;
-    remove() : Promise;
-    setPriority(priority:Priority) : Promise;
-    on(eventType:EventType, cb:((snapshot:DataSnapshot) => Promise)) : () => void;
+    setValue(value:any) : Promise<void>;
+    setValueWithPriority(value:any, priority:Priority) : Promise<void>;
+    remove() : Promise<void>;
+    setPriority(priority:Priority) : Promise<void>;
 }
 
 export type User = {
@@ -53,8 +52,8 @@ export type User = {
 
 // Description of reference received via native bridge calls
 export type DatabaseReferenceDescriptor = {
-    locationUrl: ?string;
-    key: string;
+    locationUrl?: ?string;
+    key?: string;
 }
 
 // Description of snapshot received via native bridge calls
@@ -67,8 +66,36 @@ export type DataSnapshotDescriptor = {
     priority: number;
 };
 
-type AuthCredential = {
+export type AuthCredential = {
     id: string;
     provider: string;
 };
 
+export type FacebookAuthProvider = {
+    credential(token:string) : Promise<AuthCredential>;
+};
+
+export type TwitterAuthProvider = {
+    credential(token:string, secret:string) : Promise<AuthCredential>;
+};
+
+export type GoogleAuthProvider = {
+    credential(idToken:string, accessToken:string) : Promise<AuthCredential>;
+};
+
+export type GithubAuthProvider = {
+    credential(token:string) : Promise<AuthCredential>;
+};
+
+export type AuthModule = {
+    currentUser:?User;
+    createUserWithEmail(email:string, password:string) : Promise<User>;
+    signInWithEmail(email:string, password:string) : Promise<User>;
+    signInAnonymously() : Promise<User>;
+    signInWithCredential(credential:AuthCredential|Promise<AuthCredential>) : Promise<User>;
+    FacebookAuthProvider: FacebookAuthProvider;
+    GithubAuthProvider: GithubAuthProvider;
+    TwitterAuthProvider: TwitterAuthProvider;
+    GoogleAuthProvider: GoogleAuthProvider;
+    getCurrentUser() : Promise<User>;
+};