+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Release 0.0.11
+
+  * Add currentUser
+  * Add getCurrentUser() [@Victoor #17]
+  * Fix incorrect / missing flow types
+
 ### Release 0.0.10
 
   * Add auth signOut() function